Move form content to their own components
