+++ app/settings/page.tsx
@@ -1,6 +1,6 @@
 import PageContainer from '@/components/layout/pages/page-container';
 import PageHeader from '@/components/layout/pages/page-header';
-import AccountSettings from '@/components/settings/profile-settings/account-settings';
+import AccountSettings from '@/components/settings/account-settings/account-settings';
 import ProfileSettings from '@/components/settings/profile-settings/profile-settings';
 import SideBar from '@/components/settings/settings-sidebar';
